requirement.txt
torch
torchvision
pandas
Pillow